Adding checks to limit false positives
I added some additional Regex expressions to filter out the text of the methods that lie in "quotes", @"string literals", // comments, or /* multiline comments */. There was a bit of a performance hit with these additional checks, and there's likely a more efficient way to accomplish the same checks.

I also added in some test cases for JavaScript to validate the additional checks against. This should resolve issues #5, #4, and #3.